Breaking up resonances too
