+++ python/PDFs/physics/CMakeLists.txt
@@ -24,7 +24,7 @@ goofit_add_python_docs(_Physics PDFs/physics/Amp3Body.h)
 goofit_add_python_docs(_Physics PDFs/physics/DalitzVetoPdf.h)
 goofit_add_python_docs(_Physics PDFs/physics/IncoherentSumPdf.h)
 goofit_add_python_docs(_Physics PDFs/physics/MixingTimeResolution_Aux.h)
-goofit_add_python_docs(_Physics PDFs/physics/ResonancePdf.h)
+goofit_add_python_docs(_Physics PDFs/physics/resonances/Resonance.h)
 goofit_add_python_docs(_Physics PDFs/physics/Amp3BodyT.h)
 goofit_add_python_docs(_Physics PDFs/physics/ThreeGaussResolution_Aux.h)
 goofit_add_python_docs(_Physics PDFs/physics/TruthResolution_Aux.h)
