+++ python/PDFs/physics/DalitzPlotHelpers.cpp
@@ -2,7 +2,9 @@
 
 #include <pybind11/stl.h>
 
+#include <goofit/PDFs/physics/Amplitude.h>
 #include <goofit/PDFs/physics/DalitzPlotHelpers.h>
+#include <goofit/PDFs/physics/resonances/Resonance.h>
 #include <goofit/Variable.h>
 
 using namespace GooFit;
