+++ src/PDFs/physics/SpecialDalitzIntegrator.cu
@@ -1,6 +1,7 @@
 #include <goofit/PDFs/ParameterContainer.h>
 #include <goofit/PDFs/physics/Amp3BodyT.h>
 #include <goofit/PDFs/physics/SpecialDalitzIntegrator.h>
+#include <goofit/PDFs/physics/resonances/Resonance.h>
 
 #include <thrust/transform_reduce.h>
