+++ src/PDFs/physics/SpecialResonanceCalculator.cu
@@ -1,6 +1,7 @@
 #include <goofit/PDFs/ParameterContainer.h>
 #include <goofit/PDFs/physics/DalitzPlotHelpers.h>
 #include <goofit/PDFs/physics/SpecialResonanceCalculator.h>
+#include <goofit/PDFs/physics/resonances/Resonance.h>
 
 namespace GooFit {
