+++ src/PDFs/physics/lineshapes/Flatte.cu
@@ -1,6 +1,7 @@
 #include <goofit/PDFs/physics/lineshapes/Flatte.h>
 
 #include <goofit/PDFs/ParameterContainer.h>
+#include <goofit/PDFs/physics/resonances/Resonance.h>
 
 #include "Common.h"
