+++ src/PDFs/physics/lineshapes/GLASS.cu
@@ -1,6 +1,7 @@
 #include <goofit/PDFs/physics/lineshapes/GLASS.h>
 
 #include <goofit/PDFs/ParameterContainer.h>
+#include <goofit/PDFs/physics/resonances/Resonance.h>
 
 #include "Common.h"
