+++ src/PDFs/physics/lineshapes/NonRes.cu
@@ -1,6 +1,7 @@
 #include <goofit/PDFs/physics/lineshapes/NonRes.h>
 
 #include <goofit/PDFs/ParameterContainer.h>
+#include <goofit/PDFs/physics/resonances/Resonance.h>
 
 #include "Common.h"
